requirements-common.txt
# Common shared dependencies and pinned versions
langchain==0.2.14
langchain-community==0.2.12
langchain-openai==0.1.20
openai>=1.32.0,<2.0.0
langchain-google-genai==1.0.7
google-generativeai>=0.7.0,<1.0.0
chromadb==0.5.0
pandas==2.0.2
python-dotenv==1.0.0
